fix(task_queue): check if periodic instance before delete

+++ task_queue/signals.py
@@ -20,7 +20,8 @@ def post_delete_queuetaskatnexttasks(sender, instance, **kwargs):
 
 
 def post_delete_scheduledqueuetaskgroup(sender, instance, **kwargs):
-    instance.periodic_task.delete()
+    if instance.periodic_task:
+        instance.periodic_task.delete()
 
 
 task_done = django.dispatch.Signal(providing_args=['instance'])
